feat: Adds a walking cycle
Changes the code to better visualize a walking cycle, creates a function
to figure out how to break up a distance to walk into a variable number
of steps, and fixes syntax errors in the code found in the recalculate
leg angles code. The code to find the starting leg positions had to be
adjusted as a result.

Fixes #7