feat: Adds functions to step
Adds funcitions to calculate a step and find where each leg needs to
step to. This commit was made with some debug code still attached for
testing this code at the bottom of the notebook. This commit also adds
documentation to the walk cycle function. The turn cycle cell is in
progress and is NOT a fully working cell.