feat: Adds the driver for the Lynxmotiohn SSC-32U
Creates scripts and tests for converting a servo angle to a pulse width
used in PWM and to create the grouped command for all the servos using
their desired angles. Currently all servos have the same speed. Also
adds scripts for connecting and disconnecting to the serial port and
sending the serial command to the Lynxmotiohn SSC-32U. These functions
are not tested as they also requiere the Lynxmotiohn SSC-32U.

Closes #16 issue.