perf: EMG movement fucntions step once
These functions move only take a single step and output the leg model, the
current state of which legs should move next, and the previous step so
that the funcitons can be called over and over again. Note that there
needs to be a reset function so that the stance resets before the
hexapod switches modes.