refactor: moves function for positions to movement
The code to take the walking or turning movements will be outside the
EMG based walking and turning scripts as well as the reset stance
scripts. This way all walking and turning scripts can return the foot
positions to move to and the function can iterate through them to move
the hexapod. This is in reference to issue #15