feat: Adds code to record EMG
Adds the code to record from the ADC to get the forearm EMG values.
Issue #11 still needs to be addressed to send this data wirelessly to
the other raspberry pi, but this script will record the EMG. When the
wireless communication is made, the pollEMG function will need to
incorporate it somehow.

Closes #12 issue