fix: Corrects servo smoothing
The smoothest servo movement was more from a z_resolution of 10 and not
form smoothing the servos with a previous and next servo angle. This is
because the interum postions were already found and didn't need
smoothing. The calculation done was the smoothing and I just needed to
use bigger gaps in movement.

Closes #27